+++ R/utils-interactivity.R
@@ -1,17 +1,21 @@
-## wrap clipr::clipr_available() so I can mock it
+## wrap clipr::clipr_available() so I can tweak its behaviour
 clipboard_available <- function() {
-  if (Sys.getenv("CLIPBOARD_AVAILABLE", unset = TRUE)) {
-    return(clipr::clipr_available())
+  if (interactive()) {
+    clipr::clipr_available()
+  } else {
+    isTRUE(as.logical(Sys.getenv("CLIPR_ALLOW", FALSE)))
   }
-  FALSE
 }
 
 is_testing <- function() {
   identical(Sys.getenv("TESTTHAT"), "true")
 }
 
+## this function can be replaced by rlang::is_interactive()
+## if this gets merged + released
+## https://github.com/r-lib/rlang/pull/761
 interactive <- function() {
-  base::interactive() && !is_testing()
+  rlang::is_interactive() && !is_testing()
 }
 
 ## returns TRUE if user says "no"