+++ R/utils.R
@@ -60,6 +60,18 @@ ds_is_gh <- function(venue) {
   venue
 }
 
+so_is_gh <- function(venue) {
+  if (venue == "so") {
+    message(
+      "FYI, the Stack Overflow venue \"so\" is no longer necessary. Due to ",
+      "changes at\nStack Overflow, the markdown produced by the default GitHub ",
+      "venue \"gh\" works in\nboth places. You don't need to specify it."
+    )
+    venue <- "gh"
+  }
+  venue
+}
+
 pandoc2.0 <- function() rmarkdown::pandoc_available("2.0")
 
 enfence <- function(lines,