+++ tests/testthat/test-input.R
@@ -10,7 +10,13 @@ exp_msg <- switch(
 
 test_that("reprex: clipboard input works", {
   skip_on_cran()
-  skip_if_no_clipboard()
+  # explicitly permit clipboard access in non-interactive session
+  withr::local_envvar(c(CLIPR_ALLOW = TRUE))
+  skip_if_not(
+    clipboard_available(),
+    "System clipboard is not available - skipping test."
+  )
+
   clipr::write_clip("1:5")
   expect_match(reprex(render = FALSE), "^1:5$", all = FALSE)
 })
@@ -57,9 +63,11 @@ test_that("reprex: file input in a subdirectory works", {
 
 test_that("Circular use is detected before source file written", {
   skip_on_cran()
-  ret <- reprex(y <- 2, venue = "gh", show = FALSE)
+  ret <- reprex(exp(1), venue = "gh", show = FALSE)
   expect_error(reprex(input = ret, render = FALSE), "Aborting")
-  ret <- reprex(y <- 2, venue = "so", show = FALSE)
+  ret <- reprex(exp(1), venue = "r", show = FALSE)
+  expect_error(reprex(input = ret, render = FALSE), "Aborting")
+  ret <- reprex(exp(1), venue = "html", show = FALSE)
   expect_error(reprex(input = ret, render = FALSE), "Aborting")
 })